Add a new path BFS and all-tree distance algo
Add cells_from_to to return a list of cells from start to stop, and use
this
to display the maze solution. Add distance_from to calculate the
distance
of all cells from start_cell and return the maximum distance. This is
used
in a distance (shade of blue) painter.